pyproject.toml
[project]
name = "see-to-word-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-ai-vision-imageanalysis>=1.0.0",
    "fastapi[standard]>=0.115.8",
    "python-dotenv>=1.0.1",
]
